Format and logic
Fixed formatting
Fixed logic in eq ifs
