+++ predef.py
@@ -1,11 +1,11 @@
 def prefab():
-option = str(input('INPUT TYPE : '))
-	 
-     if option == 'tg' or 'TG':
-	     print('TG D-WEIGHT IS 19LBS')
-	
-     if option == 'ug' or 'UG':
-	     print('UG D-WEIGHT IS 31LBS')
-	
-     if option == 'pg' or 'PG':
-         print('PG D-WEIGHT IS TBD')
+    option = str(input('INPUT TYPE : '))
+
+    if option in ('tg', 'TG'):
+        print('TG D-WEIGHT IS 19LBS')
+
+    elif option in ('ug', 'UG'):
+        print('UG D-WEIGHT IS 31LBS')
+
+    elif option in ('pg', 'PG'):
+        print('PG D-WEIGHT IS TBD')
